[VS-7293] restitution forms dedicated components introduced
