+++ restitution-app/ClientApp/src/app/shared/restitution/restitution-information/offender-form.component.html
@@ -8,7 +8,7 @@
     A Restitution caseworker will be in touch with you regarding the next steps.
   </p>
 
-  <h2 class="blue-header" id="ApplicantInformation" [class.mt-0]="ResitutionForm.Victim.val">{{applicant_type}} Information</h2>
+  <h2 class="blue-header mt-0" id="ApplicantInformation">{{applicant_type}} Information</h2>
   <h3>{{applicant_type}} Information</h3>
 
   <div class="row">
